+++ multiple_auth/urls.py
@@ -1,8 +1,8 @@
-from django.conf.urls import url
+from django.urls import re_path
 from .views import login, switch
 
 
 urlpatterns = [
-    url(r'^login/$', login, name='multiauth_login'),
-    url(r'^u/(?P<user_index>\d+)/$', switch, name='multiauth_switch'),
+    re_path(r'^login/$', login, name='multiauth_login'),
+    re_path(r'^u/(?P<user_index>\d+)/$', switch, name='multiauth_switch'),
 ]